custom time factor for travis
