+++ src/test/resources/application.conf
@@ -2,6 +2,7 @@ akka {
   loggers = ["de.heikoseeberger.akkalog4j.Log4jLogger"]
   loglevel = "INFO"
   test {
-    timefactor = 50.0
+    timefactor = 10
+    timefactor = ${?TEST_TIME_FACTOR}
   }
 }
